+++ openfisca_france/__init__.py
@@ -80,8 +80,8 @@ def init_country():  # drop_survey_only_variables = False, simulate_f6de = False
             # Compute one "zone APL" variable, to pre-load CSV of "code INSEE commune" to "Zone APL".
             from .model.prestations import aides_logement
             aides_logement.preload_zone_apl()
-            from .model.prelevements_obligatoires.prelevements_sociaux import taxes_salaires_main_oeuvre
-            taxes_salaires_main_oeuvre.preload_taux_versement_transport()
+            from .model.prelevements_obligatoires.prelevements_sociaux.contributions_sociales import versement_transport
+            versement_transport.preload_taux_versement_transport()
 
     return TaxBenefitSystem
